% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{d3net-package}
\alias{d3net-package}
\title{Visualizing networks using Shiny and D3.js}
\description{
Framework for creating Shiny applications for static and dynamic network
visualizations using D3.js JavaScript library.
}
\details{
The package relies on an interactive SVG graph showing the network with
force-directed placement. The vertices and and edges can be dragged around
to modify the layout. Vertex- and edge-level attributes can be shown using
vertex color and size, edge width, and dynamic tooltips.

The package enables to visualize both static network data (currently objects
of class "igraph" or "network") and dynamic network data (objects of class
"networkDynamic", currently experimental feature).

The graph is zoomable. The graph itself is also zoomable. For dynamic networks, player
controls has been provided.
}
\author{
Michal Bojanowski \email{michal2992@gmail.com},
Monika Pawluczuk \email{pawluczuk.monika@gmail.com}
}
\seealso{
\code{\link{d3net}}

\url{http://d3js.org}
}

