% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{d3net.package}
\alias{d3net.package}
\alias{d3net.package-package}
\title{Visualizing networks using d3js and Shiny}
\description{
Framework for creating Shiny applications for static and dynamic network
visualizations using d3js
}

