% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{d3net-package}
\alias{d3net-package}
\title{Visualizing networks using D3.js and Shiny}
\description{
Framework for creating Shiny applications for static and dynamic network
visualizations using D3.js JavaScript library.
}
\details{
Visualisation is a force-directed layout interactive SVG graph, which nodes and edges
can be dragged and changed dynamically to reflect network properties. The graph itself is also
zoomable. For dynamic networks, player controls has been provided.
}
\author{
Michal Bojanowski \email{michal2992@gmail.com}, Monika Pawluczuk \email{pawluczuk.monika@gmail.com}
}
\seealso{
\url{http://d3js.org}
}

