% Generated by roxygen2 (4.0.2): do not edit by hand
\name{d3net}
\alias{d3net}
\alias{d3net.default}
\alias{d3net.igraph}
\alias{d3net.network}
\alias{d3net.networkDynamic}
\title{Launching Shiny apps}
\usage{
d3net(dataset)

\method{d3net}{default}(dataset)

\method{d3net}{igraph}(dataset)

\method{d3net}{networkDynamic}(dataset)

\method{d3net}{network}(dataset)
}
\arguments{
\item{dataset}{R object, see Usage for available methods

Currently methods are implemented for objects of class "igraph", "network",
and "networkDynamic".}
}
\description{
D3net application creates a visualisation of dynamic or static network. Graph layout has some
predefined properties that can be changed through the controls in the app.
To launch the shiny app use functions described below. See Details for more information.
}
\details{
D3net application creates a visualisation of dynamic or static network with some
predefined properties. User may change these properties through
the available controls in 'd3 properties' section.
Link distance, charge and vertex size are calculated depending on total number of nodes.
To find out more about force-directed layout that was used for these visualisations,
refer to https://github.com/mbostock/d3/wiki/Force-Layout.
}
\note{
Function transforms network object into igraph object using intergraph.
}
\section{Dynamic networks}{

For dynamic visualisations, user may also set the time interval between each timepoint.
It is by default 3000ms (3s). To play with the visualisation, use the controls below the graph.
It is also possible to pause a visualisation and freeze the graph by pressing spacebar.

For visualisations with large number of vertices, nodes size can be pretty small.
For this reason, there is a zooming control that lets you change the scale.
You can also move the whole graph in every direction.
As gestures may be interferring with the dragging, it's not possible to drag the
nodes when zooming is on.

Vertices and edges properties of the visualised object can be reflected as well.
Controls in 'R properties' section enable this.
Logical and character network's vertices properties can be reflected by color nodes,
numeric ones by the size of the node. Edge properties can be reflected by edge thickness.
All of the available properties may be added to the tooltip information,
which is shown on node mouseover.
}

