% Generated by roxygen2 (4.0.1): do not edit by hand
\name{d3net}
\alias{d3net}
\alias{d3net.default}
\alias{d3net.igraph}
\alias{d3net.network}
\alias{d3net.networkDynamic}
\title{Launching Shiny apps}
\usage{
d3net(dataset)

\method{d3net}{default}(dataset)

\method{d3net}{igraph}(dataset)

\method{d3net}{networkDynamic}(dataset)

\method{d3net}{network}(dataset)
}
\arguments{
\item{dataset}{R object, see Details for available methods}
}
\description{
Functions to launch the shiny app.
}

